Extracted GitHub Usage code to multiple files
