+++ src/NuGetGallery/Views/Packages/DisplayPackage.cshtml
@@ -528,7 +528,7 @@
                 {
                     <div class="github-usage panel-collapse collapse" aria-expanded="true" id="github-usage">
                         <p>
-                            About @(Model.GitHubDependenciesInformation.TotalRepos) popular GitHub repositories depend on
+                            About @(NuGetGallery.GitHub.KiloFormatExtension.ToKiloFormat(Model.GitHubDependenciesInformation.TotalRepos)) popular GitHub repositories depend on
                             @(Model.Id), including the following:
                         </p>
 
@@ -542,7 +542,7 @@
                                             <a>@(dep.Id)</a>
                                         </span>
                                         <span class="badge pull-right" style="background-color: #004880">
-                                            @(dep.Stars.KiloFormat()) <i class="ms-Icon ms-Icon--FavoriteStarFill" aria-hidden="true" style="color:#ffac0e"></i>
+                                            @(NuGetGallery.GitHub.KiloFormatExtension.ToKiloFormat(dep.Stars)) <i class="ms-Icon ms-Icon--FavoriteStarFill" aria-hidden="true" style="color:#ffac0e"></i>
                                         </span>
                                     </button>
                                 </div>
